requirements.txt
openai
google-generativeai
PyGithub
pyyaml
